Fix launch screen configuration and UserDefaults crash
- Fixed UserDefaults crash by changing setValue:forKeyPath to set:forKey
- Fixed launch screen by properly configuring UILaunchStoryboardName in Info.plist
- Added custom LaunchScreen.storyboard with purple background and checkmark icon
- Updated app icon assets

+++ Todolist/Views/ListView.swift
@@ -12,16 +12,22 @@ struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
 
     var body: some View {
-        List {
-            ForEach(listViewModel.items) { item in
-                ListRowVIew(item: item)
-                    .onTapGesture{
-                        listViewModel.updateItem(item: item)
+        ZStack {
+            if listViewModel.items.isEmpty {
+                Text("List lis empty")
+            } else {
+                List {
+                    ForEach(listViewModel.items) { item in
+                        ListRowVIew(item: item)
+                            .onTapGesture{
+                                listViewModel.updateItem(item: item)
+                            }
                     }
-            }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
 
+                }
+            }
         }
         .listStyle(PlainListStyle())
         .navigationTitle("Todo List")
